requirements.txt
ailment==9.2.96
angr==9.2.96
angr-utils==0.5.0
arabic-reshaper==3.0.0
archinfo==9.2.96
asn1crypto==1.5.1
attrs==23.2.0
backports.strenum==1.3.1
bingraphvis==0.4.0
bitarray==2.9.2
bitstring==4.1.4
cachetools==5.3.2
capstone==5.0.0.post1
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
claripy==9.2.96
cle==9.2.96
click==8.1.7
clr-loader==0.2.6
cmd2==2.0.0
colorama==0.4.6
contourpy==1.2.0
CppHeaderParser==2.7.4
cryptography==42.0.2
cssselect2==0.7.0
cycler==0.12.1
decorator==5.1.1
dpkt==1.9.8
EasyProcess==1.1
entrypoint2==1.1
exceptiongroup==1.2.0
fonttools==4.49.0
future==0.18.3
gitdb==4.0.11
GitPython==3.1.41
graphviz==0.20.1
h11==0.14.0
html5lib==1.1
idna==3.6
itanium-demangler==1.1
Jinja2==3.1.3
kiwisolver==1.4.5
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
msgspec==0.18.6
mulpyplexer==0.9
nampa==0.1.1
networkx==3.2.1
numpy==1.26.3
oscrypto==1.3.0
outcome==1.3.0.post0
packaging==23.2
pefile==2023.2.7
peid==2.1.1
peutils==0.0.99
pillow==10.2.0
plumbum==1.8.2
ply==3.11
ppdeep==20200505
protobuf==4.25.2
psutil==5.9.8
py-tlsh==4.7.2
pycparser==2.21
pydot==2.0.0
pyelftools==0.30
pyformlang==1.0.6
Pygments==2.17.2
pyHanko==0.21.0
pyhanko-certvalidator==0.26.3
pyparsing==3.1.1
pypdf==4.0.2
pyperclip==1.8.2
pypng==0.20220715.0
PySMT==0.9.5
PySocks==1.7.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.1
pythonnet==3.0.3
pyvex==9.2.96
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.0.9
requests==2.31.0
rich==13.7.0
rpyc==5.3.1
selenium==4.18.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
svglib==1.5.1
sympy==1.12
tinycss2==1.2.1
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.10.0
tzlocal==5.2
unicorn==2.0.1.post1
unique_log_filter==0.1.0
uritools==4.0.2
urllib3==2.2.1
wcwidth==0.2.13
webdriver-manager==4.0.1
webencodings==0.5.1
wsproto==1.2.0
xhtml2pdf==0.2.15
z3-solver==4.10.2.0
